+++ src/Persistence/Behaviours/MakeUpdateRequest.php
@@ -5,7 +5,7 @@ namespace Somnambulist\Components\ApiClient\Persistence\Behaviours;
 use Psr\Log\LogLevel;
 use Somnambulist\Components\ApiClient\Client\Contracts\ConnectionInterface;
 use Somnambulist\Components\ApiClient\Persistence\Contracts\ApiActionInterface;
-use Somnambulist\Components\ApiClient\Persistence\Exceptions\EntityPersisterException;
+use Somnambulist\Components\ApiClient\Persistence\Exceptions\ActionPersisterException;
 use Symfony\Component\HttpClient\Exception\ClientException;
 use function array_values;
 use function implode;
@@ -31,7 +31,7 @@ trait MakeUpdateRequest
             $response = $this->connection->put($action->getRoute(), $action->getRouteParams(), $action->getProperties());
 
             if (200 !== $response->getStatusCode()) {
-                throw EntityPersisterException::entityNotUpdated($action->getClass(), $id, new ClientException($response));
+                throw ActionPersisterException::entityNotUpdated($action->getClass(), $id, new ClientException($response));
             }
 
             return $this->hydrateObject($response, $action->getClass());
@@ -42,7 +42,7 @@ trait MakeUpdateRequest
                 'route' => $this->connection->route($action->getRoute(), $action->getRouteParams()),
             ]);
 
-            throw EntityPersisterException::serverError($e->getMessage(), $e);
+            throw ActionPersisterException::serverError($e->getMessage(), $e);
         }
     }
 }